// src/app/financeiro/planocontas/enums/financeiro-planocontas-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 3.5.0
Code generated at time stamp: 2019-05-31T07:35:28.460
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface TipoReceitaDespesa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPlanoContaFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}